+++ cmd/natc/ippool/ipx_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) Tailscale Inc & AUTHORS
 // SPDX-License-Identifier: BSD-3-Clause
 
-package main
+package ippool
 
 import (
 	"math"
